requirements.txt
beautifulsoup4==4.12.2
celery[redis]==5.3.1
flower==2.0.0
cron-descriptor==1.4.0
psycopg2==2.9.6
Django==4.2.3
djoser==2.2.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
drf_yasg==1.21.7
django-celery-beat==2.5.0
pyTelegramBotAPI==4.13.0
annotated-types==0.5.0
selenium==4.11.2
Unidecode==1.3.6


